fix : 기본 lotto package 추가
