--- src/main/kotlin/Application.kt
@@ -1,6 +1,0 @@
-import app.DependencyInjector
-
-fun main() {
-    val injector = DependencyInjector()
-    injector.injectView()
-}
