--- src/main/kotlin/data/LottoDataSource.kt
@@ -1,5 +1,0 @@
-package data
-
-import java.util.TreeSet
-
-fun interface LottoDataSource : () -> TreeSet<Int>
