--- src/main/kotlin/domain/calculator/Calculate.kt
@@ -1,8 +1,0 @@
-package domain.calculator
-
-import domain.enums.Rank
-
-interface Calculate {
-    fun calculateWinningMoney(winning: Map<Rank, Int>): Long
-    fun calculateRateOfReturn(winningMoney: Long, purchaseLottoCount: Int): String
-}
