--- src/main/kotlin/domain/util/FormatConverter.kt
@@ -1,6 +1,0 @@
-package domain.util
-
-fun Int.convertWithDigitComma(): String = "%,d".format(this)
-
-fun Double.convertRoundAtTwoDecimal(): String = "%.1f".format(this)
-
