--- src/main/kotlin/domain/util/Retry.kt
@@ -1,11 +1,0 @@
-package domain.util
-
-fun <T> retryWhenNoException(action: () -> T): T {
-    while (true) {
-        try {
-            return action.invoke()
-        } catch (e: IllegalArgumentException) {
-            println(e)
-        }
-    }
-}
