--- src/main/kotlin/domain/validator/InputValidator.kt
@@ -1,46 +1,0 @@
-package domain.validator
-
-import domain.validator.delegate.common.CommonErrorDelegate
-import domain.validator.delegate.input.InputErrorDelegate
-import domain.enums.Output.Companion.purchaseFormat
-import domain.enums.Process
-import domain.util.ext.mapToInt
-import domain.util.ext.splitByComma
-import domain.util.ext.toMapByEachCount
-
-class InputValidator(
-    private val commonErrorDelegator: CommonErrorDelegate,
-    private val inputErrorDelegate: InputErrorDelegate
-) : InputValidate {
-
-    override fun payValidation(value: String): Pair<String, Int> {
-        val process = Process.PAY
-        commonValidation(value, process)
-        inputErrorDelegate.isThousandWonUnit(value)
-        return purchaseFormat(value)
-    }
-
-    override fun winningNumberValidation(value: String): List<Int> {
-        val process = Process.WINNING_NUMBER
-        val winningNumber = value.splitByComma()
-        commonErrorDelegator.isEmpty(value)
-        inputErrorDelegate.isInvalidInputFormat(value)
-        inputErrorDelegate.isInvalidLottoSize(winningNumber)
-        inputErrorDelegate.isExceededRange(winningNumber, process)
-        inputErrorDelegate.isDuplicated(winningNumber.toMapByEachCount())
-        return winningNumber.mapToInt()
-    }
-
-    override fun bonusNumberValidation(value: String): Int {
-        val process = Process.BONUS_NUMBER
-        commonValidation(value, process)
-        inputErrorDelegate.isExceededRange(listOf(value), process)
-        return value.toInt()
-    }
-
-    override fun commonValidation(value: String, process: Process) {
-        commonErrorDelegator.isEmpty(value)
-        commonErrorDelegator.isNumeric(value, process)
-        commonErrorDelegator.isOverIntMaxValue(value)
-    }
-}
