--- src/main/kotlin/domain/validator/delegate/common/CommonErrorDelegate.kt
@@ -1,9 +1,0 @@
-package domain.validator.delegate.common
-
-import domain.enums.Process
-
-interface CommonErrorDelegate {
-    fun isEmpty(input: String)
-    fun isNumeric(input: String, process: Process)
-    fun isOverIntMaxValue(input: String)
-}
