+++ src/main/kotlin/lotto/Application.kt
@@ -1,0 +1,8 @@
+package lotto
+
+import lotto.app.DependencyInjector
+
+fun main() {
+    val injector = DependencyInjector()
+    injector.injectView()
+}
