+++ src/main/kotlin/lotto/data/LottoDataSource.kt
@@ -1,0 +1,5 @@
+package lotto.data
+
+import java.util.TreeSet
+
+fun interface LottoDataSource : () -> TreeSet<Int>
