+++ src/main/kotlin/lotto/domain/calculator/Calculate.kt
@@ -1,0 +1,8 @@
+package lotto.domain.calculator
+
+import lotto.domain.enums.Rank
+
+interface Calculate {
+    fun calculateWinningMoney(winning: Map<Rank, Int>): Long
+    fun calculateRateOfReturn(winningMoney: Long, purchaseLottoCount: Int): String
+}
