+++ src/main/kotlin/lotto/domain/util/FormatConverter.kt
@@ -1,0 +1,6 @@
+package lotto.domain.util
+
+fun Int.convertWithDigitComma(): String = "%,d".format(this)
+
+fun Double.convertRoundAtTwoDecimal(): String = "%.1f".format(this)
+
