+++ src/main/kotlin/lotto/domain/util/Retry.kt
@@ -1,0 +1,11 @@
+package lotto.domain.util
+
+fun <T> retryWhenNoException(action: () -> T): T {
+    while (true) {
+        try {
+            return action.invoke()
+        } catch (e: IllegalArgumentException) {
+            println(e)
+        }
+    }
+}
