+++ src/main/kotlin/lotto/domain/validator/InputValidate.kt
@@ -1,0 +1,10 @@
+package lotto.domain.validator
+
+import lotto.domain.enums.Process
+
+interface InputValidate {
+    fun payValidation(value: String): Pair<String, Int>
+    fun winningNumberValidation(value: String): List<Int>
+    fun bonusNumberValidation(value: String): Int
+    fun commonValidation(value: String, process: Process)
+}
