+++ src/main/kotlin/lotto/domain/validator/delegate/common/CommonErrorDelegate.kt
@@ -1,0 +1,9 @@
+package lotto.domain.validator.delegate.common
+
+import lotto.domain.enums.Process
+
+interface CommonErrorDelegate {
+    fun isEmpty(input: String)
+    fun isNumeric(input: String, process: Process)
+    fun isOverIntMaxValue(input: String)
+}
