+++ src/test/kotlin/lotto/ApplicationTest.kt
@@ -3,7 +3,6 @@ package lotto
 import camp.nextstep.edu.missionutils.test.Assertions.assertRandomUniqueNumbersInRangeTest
 import camp.nextstep.edu.missionutils.test.Assertions.assertSimpleTest
 import camp.nextstep.edu.missionutils.test.NsTest
-import main
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
