+++ src/test/kotlin/lotto/UtilTest.kt
@@ -4,8 +4,8 @@ import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.MethodSource
-import domain.util.convertRoundAtTwoDecimal
-import domain.util.convertWithDigitComma
+import lotto.domain.util.convertRoundAtTwoDecimal
+import lotto.domain.util.convertWithDigitComma
 
 class UtilTest {
